+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2020-12-04
+
 ### Added
 
 - Config file `config.toml` + `store-path` option [#40]
@@ -230,7 +232,8 @@ First release :tada:
 
 - Linux binaries
 
-[unreleased]: https://github.com/soywod/unfog/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/soywod/unfog/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/soywod/unfog/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/soywod/unfog/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/soywod/unfog/compare/v0.4.5...v1.0.0
 [0.4.5]: https://github.com/soywod/unfog/compare/v0.4.4...v0.4.5